+++ src/games/chess/pieces/sliders/queen.rs
@@ -73,10 +73,10 @@ impl<const T: usize> Piece<T> for QueenPiece {
         &self,
         board: &Board<T>,
         from: BitBoard<T>,
-        from_bit: u32,
+        from_bit: u16,
         piece_type: usize,
-        team: u32,
-        mode: u32,
+        team: u16,
+        mode: u16,
         to: BitBoard<T>,
     ) -> BitBoard<T> {
         let lookup = self
@@ -104,8 +104,8 @@ impl<const T: usize> Piece<T> for QueenPiece {
         board: &Board<T>,
         from: BitBoard<T>,
         piece_type: usize,
-        team: u32,
-        mode: u32,
+        team: u16,
+        mode: u16,
     ) -> BitBoard<T> {
         let lookup = self
             .get_attack_lookup(board, piece_type)
